Fix the three things that only the release build was running
v0.39.0's tag build passed every gate and then failed two artifact jobs. Both
failures were in code no pull request executes, which is why they survived to a
tag.

THE SWIFT SMOKE. tests/swift/smoke.swift built a zeroed clay_stroke_preset and
handed it to clay_stroke_preset_deserialize, which has REQUIRED the caller to
declare struct_size since 0.35.0. A zeroed descriptor declares nothing, so the
call now refuses: 187/188 checks, "FAIL preset round trip". The smoke sets
struct_size on five other descriptors already; this was the one the 0.35.0
sweep missed.

CI type-checks the smoke and the release RUNS it, and the comment in ci.yml
says why: type-checking needs the header alone and costs seconds, building
against the real slices does not. That split is what let this through — the
0.35.0 break is invisible at compile time by construction (same signature, same
arity, a runtime refusal instead of a filled struct), so no type-check could
ever have caught it. v0.24.0 is the previous release this class of failure hit.

THE WINDOWS WHEEL, which builds with /WX and so reports two things clang and
gcc do not:

  grid.cpp: read_tails took palette_count as size_t and passed it to
  read_level_tail, which takes uint32_t. The value originates as a uint32_t and
  is validated at 256 or below before either sees it, so the widening was the
  anomaly rather than the narrowing. read_tails now takes uint32_t and the
  whole chain agrees.

  tape.h: a `return CLAY_TAPE_FAR;` sat after an unconditional return, left
  behind when the volume path took over the tail of that function. Dead, and
  deleted rather than silenced.

Verified: 188/188 swift smoke checks on both slices, where it was 187/188;
kernel dialect OK across cpu+cuda+metal for all 15 headers; 5/5 ctest.

+++ src/voxel/grid.cpp
@@ -1192,7 +1192,7 @@ std::optional<VoxelGrid> VoxelGrid::deserialize(const std::uint8_t* data, std::s
 // which is what every tail here is an addition to. False means a tail WAS
 // recognised and then failed to decode, which is a corrupt file.
 bool VoxelGrid::read_tails(const std::uint8_t* data, std::size_t size, std::size_t* pos,
-                           std::size_t palette_count) {
+                           std::uint32_t palette_count) {
     auto tag_at = [&](std::uint32_t* out) {
         if (*pos + 4 > size) return false;
         *out = 0;
